(stage) Борьба с vallgrid

+++ likhachev.aleksey/T1/main.cpp
@@ -52,7 +52,7 @@ int main()
 
   if (shapeCount) {
     std::cout << shapes[0]->getArea();
-    for (int i = 0; i < shapeCount; i++) {
+    for (size_t i = 0; i < shapeCount; i++) {
       delete[] shapes[i];
     }
   }
